backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pypdf>=5.0.0
cryptography>=3.1
python-multipart>=0.0.9
